Move store and config folders inside src
